I've added a parallel module to get the number of threads and also to create a set of seeded random number generators (one for each thread)
The recovery loop works, using a different generator for each thread.
This gives the same answer as C for one thread, and goes twice
as quickly for two threads. Learned that I have to save values
to an array / memory view, as every other variable that is assigned
is treated as private to the loop